full functionality

+++ HW1/Credit.java
@@ -1,17 +1,18 @@
+import java.math.BigInteger;
 import java.util.Scanner;
 
 public class Credit implements TransactionBehavior{
     public float totalDue;
     public float amountPaid;
     public float changeDue;
-    public int cardNumber;
+    public BigInteger cardNumber;
 
     public void transaction(){
 
         Scanner input = new Scanner(System.in);
 
         System.out.print("Enter card number: ");
-        cardNumber = input.nextInt();
+        cardNumber = input.nextBigInteger();
         
         System.out.print("Enter amount due: ");
         totalDue = input.nextFloat();
@@ -20,7 +21,6 @@ public class Credit implements TransactionBehavior{
 
         System.out.println("**************************************");
 
-        //input.close();
 
     }
